made nearby reports more modular
